+++ app/src/main/AndroidManifest.xml
@@ -20,6 +20,7 @@
         <meta-data
             android:name="com.naver.maps.map.CLIENT_ID"
             android:value="6u9a727lyo" />
+        <activity android:name=".MapFragmentActivity"/>
     </application>
 
 </manifest>
